rename TokenUtils in tests
